# getnet-cli-nuek/src/main/resources/api/api-definition-0.0.1.yml
openapi: 3.0.3
info:
  title: Merchant Portal (Converted from Postman)
  version: "1.0.0"
servers:
  - url: https://srvnuarintra.santander.pru.bsch
  - url: https://repgtnprep.minsaitpaymentsprocessing.com
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    TpvOperationsResponse:
      type: object
      properties:
        numOperations:
          type: integer
          format: int32
        clearances:
          type: array
          items:
            $ref: "#/components/schemas/TpvOperation"
...
    TpvOperation:
      properties:
        amountDiscounted:
          type: number
          format: double
          example: 0.12
        amountSubscribed:
          type: number
          format: double
          example: 34.88
        amountTotal:
          type: number
          format: double
          example: 35.00,
        refTNumOpe:
          type: string
          example: ' 000060'
        returns:
          type: integer
          format: int32
          example: 0
        sales:
          type: integer
          format: int32
          example: 22
        totalDate:
          type: string
          example: '2025-10-24'
        totalOrder:
          type: integer
          format: int32
          example: 680444
        totalTPV:
          type: string
          example: '1'
        totalTime:
          type: string
          example: '20:30'
        valueDate:
          type: string
          example: '2025-10-27'
        cardType:
          type: string
          example: 'VISA'
        currency:
          type: string
          example: 'EUR'
...
paths:
...
  /api/TPV/tpv_clearance:
    get:
      operationId: getTpvOperations
      summary: NUek – TPV Operations
      tags: [Nuek]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: commerceContract
          required: true
          schema:
            type: string
            example: '004918125240196905'
        - in: query
          name: order
          required: true
          schema:
            type: string
            example: 'D'
        - in: query
          name: dateFrom
          required: true
          schema:
            type: string
            example: '20251001'
        - in: query
          name: dateTo
          required: true
          schema:
            type: string
            example: '20251024'
      responses:
        "200":
          description: "List of TPV clearance operations"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TpvOperationsResponse"



  ###########################################################
  ## 5. NUek – Transacciones TPV
  ###########################################################
...